+++ app/src/main/java/com/example/nostack/model/User/Attendee.java
@@ -1,6 +1,6 @@
-package com.example.nostack;
+package com.example.nostack.model.User;
 
-public class Attendee extends User{
+public class Attendee extends User {
     public Attendee(String firstName, String lastName, String username, String emailAddress, String phoneNumber, String uuid) {
         super(firstName, lastName, username, emailAddress, phoneNumber, uuid);
         setRole("Attendee");
