+++ app/src/main/java/com/example/nostack/model/User/User.java
@@ -1,4 +1,4 @@
-package com.example.nostack;
+package com.example.nostack.model.User;
 
 /**
  * User will
